// src/app/api/generated/model/buildingDTO.ts
/**
 * Hausverwaltung API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BuildingDTO { 
    id: number;
    name: string;
    postCode: number;
    city: string;
    address: string;
    contactPerson?: string;
    contactPhone?: string;
    buildingType?: BuildingDTOBuildingTypeEnum;
    active?: boolean;
}
export enum BuildingDTOBuildingTypeEnum {
    Residential = 'RESIDENTIAL',
    Office = 'OFFICE',
    Industrial = 'INDUSTRIAL'
};